[#58] Remove @app-utils alias and fix eslint for import paths [#58] Rename script for linting
