+++ app/src/app/datasets/_partials/dataset-listing.tsx
@@ -3,7 +3,9 @@
 import { QueryKey, useQuery } from '@app-api'
 import { EmptyState, LayoutPane, LoadingSpinner } from '@app-components'
 import { useTranslation } from '@app-i18n'
-import { formatNumber, useLocalStorage, useVirtualScroll, type VirtualItem } from '@app-utils'
+import { formatNumber } from '@app/shared/utils/js-utils'
+import { useLocalStorage } from '@app/shared/utils/use-local-storage'
+import { useVirtualScroll, type VirtualItem } from '@app/shared/utils/use-virtual-scroll'
 import { SearchSvg, TextField, TOKENS__SPACING } from '@ds/core'
 import { useSearchParams } from 'next/navigation'
 import { useMemo, useState } from 'react'
