+++ app/src/app/datasets/_partials/dataset-preview.tsx
@@ -3,7 +3,9 @@
 import { QueryKey, useQuery } from '@app-api'
 import { DataTable, EmptyState, LayoutPane, LoadingSpinner } from '@app-components'
 import { useCountries, useTranslation } from '@app-i18n'
-import { convertJsonStatToTable, formatDate, formatNumber, type TableRowValue, useLocalStorage } from '@app-utils'
+import { formatDate, formatNumber } from '@app/shared/utils/js-utils'
+import { convertJsonStatToTable, type TableRowValue } from '@app/shared/utils/json-stat'
+import { useLocalStorage } from '@app/shared/utils/use-local-storage'
 import { useSearchParams } from 'next/navigation'
 import { type ReactNode, useEffect, useMemo } from 'react'
 import { EurostatApi } from '../_api/eurostat-api'
