+++ app/src/shared/components/data-table/data-table.stories.tsx
@@ -1,4 +1,4 @@
-import { type TableCol } from '@app-utils'
+import { type TableCol } from '@app/shared/utils/json-stat'
 import { defineMeta, loremArray, loremInt, loremText } from '@ds/docs/core'
 import { type Meta, type StoryObj } from '@storybook/nextjs-vite'
 import { DataTable } from './data-table'
