--- app/src/shared/utils/index.ts
@@ -1,4 +1,0 @@
-export * from './_js-utils'
-export * from './_json-stat'
-export * from './_use-local-storage'
-export * from './_use-virtual-scroll'
