+++ app/src/shared/utils/js-utils.ts
