+++ app/src/shared/utils/json-stat.ts
