+++ app/tsconfig-sb.json
@@ -10,8 +10,8 @@
 
     /* Modules & Resolution */
     "paths": {
+      // ESLint doesn't detect "@app-utils/*.ts" imports
       "@app-components": ["./src/shared/components/index.ts"],
-      "@app-utils": ["./src/shared/utils/index.ts"],
       "@app-i18n": ["./src/core/i18n/index.ts"],
       "@app-api": ["./src/core/api/index.ts"],
       "@app/*": ["./src/*"],
